All components are now coupled & compiled
Driving component still needs to have its logic reviewed and tested, but otherwise we're good to go. ObstacleDetectionComponent now now writes on screen whether to go or not. In order for it to stop the car, however, the logic in the steering controller and driving controller still needs to be written.